refactor mongo utils
